Make tests work on Windows
Tests were not working on Windows for three reasons:
1. We have a abs macro in astonia.h that was meant to provide abs value if a platform doesn't provide it already. Today all platforms provide this, so it's redundant and was clobbering the standard library absolute value on Windows in tests, in a way that was not compatible
2. Removing that exposed a real bug in absolute value calculations that is now fixed
3. On Windows only SDL redefines main and SDL_main to run Windows specific init logic that isn't needed on OSX and Linux. This update the signature of main in tests to match that of SDL_main, so the override works as expected on Windows
4. SDL libraries need to be included in a different way on Windows in the test Makefile - now we do platform detection and include them correctly.

Tested and working in MSYS2 shell and CMD.exe. Does not work on WSL, but honestly, if you are on Windows and can't figure out how to use either CMD.exe or MSYS2, I'm not sure we should overcomplicate things just so this also works on WSL...

+++ src/game/sprite.c
@@ -4023,7 +4023,7 @@ DLL_EXPORT unsigned int _trans_asprite(map_index_t mn, unsigned int sprite, tick
 	case 59195:
 		c2 = IRGB(
 		    abs(30 - (int)(attick % 61)) / 4, abs(30 - (int)(attick % 61)) / 2 + 5, abs(30 - (int)(attick % 61)) / 4);
-		light = abs(30 - (attick % 61)) / 3;
+		light = abs(30 - (int)(attick % 61)) / 3;
 		shine = 5;
 		sprite = 51085;
 		break;
